-- ACBrTEFComum -- [-] Ajuste em TACBrTEFLinha, mudando a propruedade "Identificacao" de SmallInt     para Integer
git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@21986 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrTEFD/ACBrTEFComum.pas
@@ -83,7 +83,7 @@ type
 
   TACBrTEFLinha = class
   private
-    FIdentificacao: smallint;
+    FIdentificacao: Integer;
     FACBrTEFDLinhaInformacao: TACBrInformacao;
     FLinha: String;
     FSequencia: smallint;
@@ -97,7 +97,7 @@ type
 
     property Linha: String read GetLinha write SetLinha;
 
-    property Identificacao: smallint read FIdentificacao;
+    property Identificacao: Integer read FIdentificacao;
     property Sequencia: smallint read FSequencia;
     property Chave: String read GetChave;
     property Informacao: TACBrInformacao read FACBrTEFDLinhaInformacao;
